+++ EQWOWConverter/Files/WOWFiles/SQL/World/CreatureSQL.cs
@@ -24,7 +24,8 @@ namespace EQWOWConverter.WOWFiles
         public override string DeleteRowSQL()
         {
             StringBuilder sb = new StringBuilder();
-            sb.AppendLine("DELETE FROM creature WHERE `guid` >= " + Configuration.SQL_CREATURE_GUID_DEBUG_LOW.ToString() + " AND `guid` <= " + Configuration.SQL_CREATURE_GUID_DEBUG_HIGH + ";");
+            if (Configuration.CREATURE_SPAWN_AND_WAYPOINT_DEBUG_MODE == true)
+                sb.AppendLine("DELETE FROM creature WHERE `guid` >= " + Configuration.SQL_CREATURE_GUID_DEBUG_LOW.ToString() + " AND `guid` <= " + Configuration.SQL_CREATURE_GUID_DEBUG_HIGH + ";");
             sb.Append("DELETE FROM creature WHERE `guid` >= " + Configuration.SQL_CREATURE_GUID_LOW.ToString() + " AND `guid` <= " + Configuration.SQL_CREATURE_GUID_HIGH + ";");          
             return sb.ToString();
         }
